+++ bin/src/mock_testnet.rs
@@ -1,11 +1,11 @@
 use anyhow::Result;
 use ethers_providers::{Http, Provider};
 use itertools::Itertools;
-use prover::circuit::{
+use prover::zkevm::circuit::{
     block_traces_to_witness_block, calculate_row_usage_of_witness_block, SuperCircuit,
     SUB_CIRCUIT_NAMES,
 };
-use prover::prover::Prover;
+use prover::zkevm::Prover;
 use prover::utils::init_env_and_log;
 use reqwest::Url;
 use serde::Deserialize;
