+++ ffi/src/prove.rs
@@ -1,11 +1,11 @@
 use crate::utils::{c_char_to_str, c_char_to_vec, vec_to_c_char};
 use libc::c_char;
-use prover::prover::Prover;
+use prover::zkevm;
 use prover::utils::init_env_and_log;
 use std::cell::OnceCell;
 use types::eth::BlockTrace;
 
-static mut PROVER: OnceCell<Prover> = OnceCell::new();
+static mut PROVER: OnceCell<zkevm::Prover> = OnceCell::new();
 
 /// # Safety
 #[no_mangle]
@@ -13,7 +13,7 @@ pub unsafe extern "C" fn init_prover(params_path: *const c_char, _seed_path: *co
     init_env_and_log("ffi_prove");
 
     let params_path = c_char_to_str(params_path);
-    let p = Prover::from_param_dir(params_path);
+    let p = zkevm::Prover::from_param_dir(params_path);
     PROVER.set(p).unwrap();
 }
