+++ prover/src/zkevm.rs
@@ -1,0 +1,8 @@
+mod capacity_checker;
+pub mod circuit;
+pub mod prover;
+pub mod verifier;
+
+pub use capacity_checker::CircuitCapacityChecker;
+pub use prover::{Prover, AggCircuitProof, TargetCircuitProof};
+pub use verifier::Verifier;
