+++ prover/src/zkevm/capacity_checker.rs
@@ -3,7 +3,7 @@ use mpt_zktrie::state::ZktrieState;
 use serde_derive::{Deserialize, Serialize};
 use types::eth::BlockTrace;
 
-use crate::circuit::{
+use super::circuit::{
     block_traces_to_witness_block_with_updated_state, calculate_row_usage_of_witness_block,
     update_state, DEGREE, SUB_CIRCUIT_NAMES,
 };
