+++ prover/src/zkevm/prover/mock.rs
@@ -1,5 +1,5 @@
+use super::super::circuit::{block_traces_to_witness_block, check_batch_capacity, TargetCircuit, DEGREE};
 use super::Prover;
-use crate::circuit::{block_traces_to_witness_block, check_batch_capacity, TargetCircuit, DEGREE};
 use crate::utils::metric_of_witness_block;
 use anyhow::bail;
 use halo2_proofs::dev::MockProver;
