+++ prover/src/zkevm/prover/util.rs
@@ -1,7 +1,7 @@
 //! Initialization and utility APIs for Prover.
 //!
+use super::super::circuit::{TargetCircuit, AGG_DEGREE, DEGREE};
 use super::Prover;
-use crate::circuit::{TargetCircuit, AGG_DEGREE, DEGREE};
 use crate::utils::{load_params, DEFAULT_SERDE_FORMAT};
 use halo2_proofs::halo2curves::bn256::Bn256;
 use halo2_proofs::plonk::keygen_pk2;
