+++ prover/tests/integration.rs
@@ -1,15 +1,12 @@
 use chrono::Utc;
 use halo2_proofs::{plonk::keygen_vk, SerdeFormat};
+use prover::zkevm::{CircuitCapacityChecker, Prover, Verifier};
+use prover::zkevm::circuit::{SuperCircuit, TargetCircuit, DEGREE};
 use prover::{
-    capacity_checker::CircuitCapacityChecker,
-    circuit::{SuperCircuit, TargetCircuit, DEGREE},
     io::serialize_vk,
-    prover::Prover,
     utils::{get_block_trace_from_file, init_env_and_log, load_or_create_params, load_params},
 };
-
-use prover::test_util;
-use test_util::{load_block_traces_for_test, PARAMS_DIR};
+use prover::test_util::{load_block_traces_for_test, PARAMS_DIR};
 
 use zkevm_circuits::util::SubCircuit;
 
@@ -78,27 +75,22 @@ fn test_capacity_checker() {
 
 #[test]
 fn estimate_circuit_rows() {
-    use prover::circuit::{self, TargetCircuit};
 
     init_env_and_log("integration");
 
     let (_, block_trace) = load_block_traces_for_test();
 
     log::info!("estimating used rows for batch");
-    let rows = circuit::SuperCircuit::estimate_rows(&block_trace);
+    let rows = SuperCircuit::estimate_rows(&block_trace);
     log::info!("super circuit: {:?}", rows);
 }
 
 #[cfg(feature = "prove_verify")]
 #[test]
 fn test_mock_prove() {
-    use prover::circuit;
-
-    use crate::test_util::load_block_traces_for_test;
-
     init_env_and_log("integration");
     let block_traces = load_block_traces_for_test().1;
-    Prover::mock_prove_target_circuit_batch::<circuit::SuperCircuit>(&block_traces).unwrap();
+    Prover::mock_prove_target_circuit_batch::<SuperCircuit>(&block_traces).unwrap();
 }
 
 #[cfg(feature = "prove_verify")]
@@ -211,8 +203,6 @@ fn test_vk_same() {
 
 fn test_target_circuit_prove_verify<C: TargetCircuit>() {
     use std::time::Instant;
-
-    use prover::verifier::Verifier;
 
     init_env_and_log("integration");
